Create copy service
