+++ gse-app/pom.xml
@@ -62,7 +62,7 @@
         </dependency>
         <dependency>
             <groupId>com.powsybl</groupId>
-            <artifactId>powsybl-gse-copypaste-afs</artifactId>
+            <artifactId>powsybl-gse-copy-paste-afs</artifactId>
             <version>${project.version}</version>
         </dependency>
         <!-- Test dependencies -->
