--- gse-copypaste/gse-copypaste-afs-server/src/main/java/com/powsybl/gse/copypaste/afs/server/CopyApplication.java
@@ -1,8 +1,0 @@
-package com.powsybl.gse.copypaste.afs.server;
-
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
-
-@ApplicationPath("/rest")
-public class CopyApplication extends Application {
-}
