--- gse-copypaste/gse-copypaste-afs/src/main/java/com/powsybl/gse/copypaste/afs/copyExceptions/CopyPasteDifferentFileSystemNameException.java
@@ -1,10 +1,0 @@
-package com.powsybl.gse.copypaste.afs.copyExceptions;
-
-import com.powsybl.gse.copypaste.afs.CopyPasteException;
-
-public class CopyPasteDifferentFileSystemNameException extends CopyPasteException {
-
-    public CopyPasteDifferentFileSystemNameException(String msg) {
-        super(msg);
-    }
-}
