--- gse-copypaste/gse-copypaste-afs/src/main/resources/lang/CopyManager.properties
@@ -1,3 +1,0 @@
-Copy=Copy
-CopyError=Copy Error
-FileExists=A file with the name already exists in this folder
