--- gse-copypaste/gse-copypaste-afs/src/main/resources/lang/CopyManager_fr.properties
@@ -1,3 +1,0 @@
-Copy=Copie
-CopyError=Erreur de copie
-FileExists=un fichier de mème nom existe déja dans ce dossier
