--- gse-copypaste/pom.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2019, RTE (http://www.rte-france.com)
-    This Source Code Form is subject to the terms of the Mozilla Public
-    License, v. 2.0. If a copy of the MPL was not distributed with this
-    file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>com.powsybl</groupId>
-        <artifactId>powsybl-gse</artifactId>
-        <version>1.7.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>powsybl-gse-copypaste</artifactId>
-
-    <packaging>pom</packaging>
-
-    <modules>
-        <module>gse-copypaste-afs</module>
-        <module>gse-copypaste-afs-local</module>
-        <module>gse-copypaste-afs-remote</module>
-        <module>gse-copypaste-afs-server</module>
-    </modules>
-
-
-</project>
