+++ gse-util/pom.xml
@@ -70,8 +70,8 @@
 
         <dependency>
             <groupId>com.powsybl</groupId>
-            <artifactId>powsybl-gse-copypaste-afs</artifactId>
-            <version>1.7.0-SNAPSHOT</version>
+            <artifactId>powsybl-gse-copy-paste-afs</artifactId>
+            <version>${project.version}</version>
         </dependency>
     </dependencies>
 </project>
